refactor(components): split preview and source-view into directories
Mirror the file-tree/ layout for the remaining components so each one is
a directory with index.tsx as the entry, sub-components and pure helpers
as siblings, and tests co-located with the file they test.

preview/
  index.tsx          - Preview wrapper
  zoom-controls.tsx  - extracted from the inline ZoomControls in preview.tsx
  zoom.ts            - MIN_SCALE / MAX_SCALE / clampScale
  use-keyboard-pan.* - moved in from the loose components/ sibling

source-view/
  index.tsx          - SourceView
  line-row.tsx       - extracted per-line row with the fold gutter button
  highlighter.ts     - lazy shiki singleton + highlight() wrapper
  source-fold.*      - moved in from the loose components/ sibling
