+++ frontend/src/components/preview/use-keyboard-pan.ts
